added api endpoint

+++ application/source/source/presentation_layer/data_transfer_object/request/_in_context_for/presentation_layer/service/actix_web_component/request_handler/application_programming_interface/version_1/mobile/_in_context_for/domain_layer/entity/entity/channel/_new_for_context/base/_new_for_context/get_many_by_subscribers_quantity/base/mod.rs
@@ -3,15 +3,15 @@ use serde::Deserialize;
 #[derive(Deserialize)]
 pub struct Base {
     #[serde(rename = "csq")]
-    channel_subscribers_quantity: Option<u64>,
+    channel_subscribers_quantity: Option<i64>,
     #[serde(rename = "o")]
-    order: u8,
+    order: i8,
     #[serde(rename = "l")]
-    limit: u8
+    limit: i16
 }
 
 impl Base {
-    pub fn into_inner(self) -> (Option<u64>, u8, u8) {
+    pub fn into_inner(self) -> (Option<i64>, i8, i16) {
         return (
             self.channel_subscribers_quantity,
             self.order,
